% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrockbio-package.R
\docType{package}
\name{bedrockbio-package}
\alias{bedrockbio}
\alias{bedrockbio-package}
\title{bedrockbio: Open-Access Computational Biology Datasets}
\description{
Efficiently access the 'Bedrock Bio' library of open-access computational biology datasets. Lazily query datasets backed by 'DuckDB' and 'Apache Iceberg', with support for predicate pushdown and column projection to the cloud storage backend. This enables quick, iterative access to otherwise massive, unwieldy datasets without downloading them in full. See \url{https://bedrock.bio} for available datasets and documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bedrock.bio}
  \item \url{https://github.com/bedrock-bio/bedrock-bio-client}
  \item Report bugs at \url{https://github.com/bedrock-bio/bedrock-bio-client/issues}
}

}
\author{
\strong{Maintainer}: Liam Abbott \email{liam@bedrock.bio} [copyright holder]

}
\keyword{internal}
