% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List available tables in the Bedrock Bio library}
\usage{
list_tables()
}
\value{
A character vector of table identifiers
}
\description{
List available tables in the Bedrock Bio library
}
\examples{
\dontrun{
library(bedrockbio)
list_tables()
}

}
